Fix flash_attn import check and text_encoder dtype parameter
- is_flash_attn_available() now verifies actual importability of flash_attn_func
  instead of just checking module existence, avoiding errors with old versions.
- Use torch_dtype instead of dtype for AutoModel.from_pretrained() to match
  the transformers API.

+++ src/utils/loader.py
@@ -178,8 +178,7 @@ def load_from_local_dir(
     text_encoder_dir = model_dir / "text_encoder"
     text_encoder = AutoModel.from_pretrained(
         str(text_encoder_dir),
-        # torch_dtype=dtype, # some version use this
-        dtype=dtype,
+        torch_dtype=dtype,
         trust_remote_code=True,
     )
     text_encoder.to(device)
